Renamed `Resources` to `resources`
