+++ resources/egg_group.py
