+++ utils.py
@@ -5,7 +5,7 @@ import re
 
 import requests
 
-from Resources.data import AbstractData
+from resources.data import AbstractData
 from console import console
 from readchar import readkey, key as keys
